Feat/notifications

+++ _backend/apps/strava-service/src/strava-webhook/strava-webhook.processor.ts
@@ -3,7 +3,7 @@ import { Job, Queue } from 'bullmq';
 import { InjectPinoLogger, PinoLogger } from 'nestjs-pino';
 import { StravaWebhookEventDto } from './dto/strava-webhook-event.dto';
 import { StravaWebhookService } from './strava-webhook.service';
-import stravaJsonData_forTest from '../example_strava_data.json';
+import stravaJsonData_forTest from '../example_strava_data/example_strava_data.json';
 
 // Decorator name must exactly match the Queue name registered in the module
 @Processor('strava-webhook-queue')
@@ -17,11 +17,16 @@ export class StravaProcessor extends WorkerHost {
   }
   // NestJS automatically calls this method when a new job appears in the queue
   async process(job: Job<any, any, string>) {
-    this.logger.info({ custom: true }, `Worker starting job: ${job.id}`);
+    if (job.name === 'test-strava-event') {
+      console.log(`Processing TEST MOCK STRAVA DATA id: ${job.id}, name: ${job.name}`);
+      const slimActivityData = this.stravaWebhookService.simplifyActivityData(this.getRandomMockStravaActivityData());
+      return { type: 'test', data: slimActivityData };
+    }
+
     if (job.name === 'process-strava-event') {
       const stravaEvent: StravaWebhookEventDto = job.data;
 
-      // ------------ Delete ------------
+      // ------------------------ Delete ------------------------
       if (stravaEvent.aspect_type === 'delete') {
         await this.stravaWebhookService.deleteActivityData(stravaEvent.owner_id, stravaEvent.object_id);
         return { type: 'delete', data: stravaEvent };
@@ -29,16 +34,10 @@ export class StravaProcessor extends WorkerHost {
       this.logger.debug({ custom: true }, `Processing strava event: ${JSON.stringify(stravaEvent)}`);
 
       // Fetch activity from Strava API
-      let activityData = stravaJsonData_forTest;
-      if ('test' in job.data) {
-        // ------- DELETE THIS IN PRODUCTION -------
-        // ----- TESTING STRAVA JSON DATA -----
-        activityData = stravaJsonData_forTest;
-        // ------- DELETE THIS IN PRODUCTION -------
-
-        // throw new Error('Test error: Simulating failure in fetching activity data');
-      } else
-        activityData = await this.stravaWebhookService.downloadActivity(stravaEvent.object_id, stravaEvent.owner_id);
+      const activityData = await this.stravaWebhookService.downloadActivity(
+        stravaEvent.object_id,
+        stravaEvent.owner_id,
+      );
       this.logger.debug({ custom: true }, `Gear: ${JSON.stringify(activityData.data.gear)}`);
 
       // If activity is not Ride or EBikeRide, skip processing
@@ -48,7 +47,7 @@ export class StravaProcessor extends WorkerHost {
       const slimActivityData = this.stravaWebhookService.simplifyActivityData(activityData.data);
 
       // Decision based on type activity create / update / delete
-      // ------------ Create ------------
+      // ------------------------ Create ------------------------
       if (stravaEvent.aspect_type === 'create') {
         await this.stravaWebhookService.saveActivityData(
           activityData.data,
@@ -78,7 +77,7 @@ export class StravaProcessor extends WorkerHost {
       return;
     }
     const jobName =
-      result.type === 'create'
+      result.type === 'create' || result.type === 'test'
         ? 'strava_activity-created'
         : result.type === 'update'
           ? 'strava_activity-updated'
@@ -92,5 +91,16 @@ export class StravaProcessor extends WorkerHost {
   @OnWorkerEvent('failed')
   onFailed(job: Job, error: Error): void {
     this.logger.error({ err: error.message, jobId: job.id }, 'Job failed');
+  }
+  private getRandomMockStravaActivityData(): typeof stravaJsonData_forTest.data {
+    const mockData = structuredClone(stravaJsonData_forTest);
+
+    mockData.data.id = Math.floor(Math.random() * 1_000_000_000);
+    mockData.data.start_date = new Date(Date.now() - Math.floor(Math.random() * 1_000_000_000)).toISOString();
+    mockData.data.distance = Math.floor(Math.random() * 200) * 10000;
+    console.log(mockData.data.distance);
+    mockData.data.moving_time = Math.floor(Math.random() * 100000);
+
+    return mockData.data;
   }
 }
